requirements.txt
fastapi==0.104.1
solana==0.30.2
fpdf2==2.7.6
uvicorn[standard]==0.24.0
python-multipart==0.0.6